refactor: Move logging/reporting e2e tests to `/test` directory
