+++ test/logging/testdata/config-with-cycle/manifest.yaml
