+++ test/logging/testdata/config-with-cycle/project/reports/report.json
