+++ test/logging/testdata/configs-with-duplicate-ids/project/alerting-profile/another_config.yaml
@@ -1,0 +1,8 @@
+configs:
+- id: profile
+  config:
+    name: Star Gate Service
+    template: profile.json
+    skip: false
+  type:
+    api: alerting-profile
