+++ test/logging/testdata/configs-with-duplicate-ids/project/auto-tag/config.yaml
@@ -1,0 +1,13 @@
+configs:
+- id: tag # this does not overlap with project2/auto-tag/tag
+  config:
+    name: Resistance is futile!
+    template: application-tagging.json
+  type:
+    api: auto-tag
+- id: profile # this does not overlap with project/alerting-profile/profile configs
+  config:
+    name: Resistance is futile!
+    template: application-tagging.json
+  type:
+    api: auto-tag
