+++ test/logging/testdata/key-user-action-without-scope/manifest.yaml
