+++ test/logging/testdata/key-user-action-without-scope/project/key-user-actions-web/kuaweb.json
