+++ test/logging/testdata/references/classic-with-settings-mngt-zone/profile.json
@@ -1,0 +1,76 @@
+{
+    "displayName": "{{ .name }}",
+    "rules": [
+        {
+            "severityLevel": "AVAILABILITY",
+            "tagFilter": {
+                "includeMode": "INCLUDE_ALL",
+                "tagFilters": [
+                    {
+                        "context": "CONTEXTLESS",
+                        "key": "{{ .name }}",
+                        "value": null
+                    }
+                ]
+            },
+            "delayInMinutes": 0
+        },
+        {
+            "severityLevel": "ERROR",
+            "tagFilter": {
+                "includeMode": "INCLUDE_ALL",
+                "tagFilters": [
+                    {
+                        "context": "CONTEXTLESS",
+                        "key": "{{ .name }}",
+                        "value": null
+                    }
+                ]
+            },
+            "delayInMinutes": 0
+        },
+        {
+            "severityLevel": "PERFORMANCE",
+            "tagFilter": {
+                "includeMode": "INCLUDE_ALL",
+                "tagFilters": [
+                    {
+                        "context": "CONTEXTLESS",
+                        "key": "{{ .name }}",
+                        "value": null
+                    }
+                ]
+            },
+            "delayInMinutes": 0
+        },
+        {
+            "severityLevel": "RESOURCE_CONTENTION",
+            "tagFilter": {
+                "includeMode": "INCLUDE_ALL",
+                "tagFilters": [
+                    {
+                        "context": "CONTEXTLESS",
+                        "key": "{{ .name }}",
+                        "value": null
+                    }
+                ]
+            },
+            "delayInMinutes": 0
+        },
+        {
+            "severityLevel": "MONITORING_UNAVAILABLE",
+            "tagFilter": {
+                "includeMode": "INCLUDE_ALL",
+                "tagFilters": [
+                    {
+                        "context": "CONTEXTLESS",
+                        "key": "{{ .name }}",
+                        "value": null
+                    }
+                ]
+            },
+            "delayInMinutes": 0
+        }
+    ],
+    "managementZoneId": "{{ .managementZoneId }}"
+}
