+++ test/logging/testdata/references/invalid-classic-with-settings/config.yaml
@@ -1,0 +1,32 @@
+configs:
+  - id: profile
+    type:
+      settings:
+        schema: builtin:alerting.profile
+        scope: environment
+    config:
+      name: profile-icws
+      template: profile.json
+      parameters:
+        managementZoneId: [ builtin:management-zones, zone, id ]
+  - id: zone
+    type:
+      settings:
+        schema: builtin:management-zones
+        scope: environment
+    config:
+      name: zone-icws
+      parameters:
+        environment: environment1
+        meId: HOST_GROUP-1234567890123456
+      template: zone.json
+  - id: slack
+    type:
+      api: notification
+    config:
+      name: notification-icws
+      parameters:
+        alertingProfileId: [builtin:alerting.profile, profile, id]
+        environment: Env1
+      template: slack.json
+      skip: false
