+++ test/logging/testdata/support-archive/invalid-manifest.yaml
