+++ test/logging/testdata/support-archive/manifest.yaml
